Handle multiple TextEditor for the same content
Make the message cache global to the entire package. When a new
TextEditor Minimap is registered check whether there are any existing
messages in the cache for it, if any were found mark them in this
instance.

Note that currently Minimap triggers this before the TextEditor is
actually visible, leading to many cascading bugs. Defer the marking of
these old messages to an idle callback to give the TextEditor time to
initialize. This needs to be fixed in Minimap itself to generalize this
to all plugins.

Fixes #16.