Add some sanity checks and prevent a memory leak
When adding existing messages to a new binding in the delayed callback,
make sure the message hasn't already been sent to the binding.

When adding a new message in a binding, check that it hasn't already
been marked.

Remove references to decorations that have been destroyed already so we
aren't holding a reference to them, allowing the DisplayMarkerLayer to
do with them as it pleases.